requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
pydantic>=2.0.0
sqlalchemy>=2.0.0
alembic>=1.10.0
aiohttp>=3.8.4
tenacity>=8.2.2
python-dotenv>=1.0.0
loguru>=0.7.0
google-generativeai>=0.3.0
openai>=1.0.0
anthropic>=0.5.0
langchain>=0.0.267
langchain-community>=0.0.10
qdrant-client>=1.5.0
sentence-transformers>=2.2.2
pillow>=10.0.0
pytest>=7.3.1
pytest-asyncio>=0.21.0
httpx>=0.24.0
asgi-lifespan>=2.1.0
